Add files via upload
These files were produced to be implemented on a microcontroller (TI F280027 LaunchPad) in order to be able to control a group of PE converters so that Green Hydrogen could be produced. 

This was applied to a real-life prototype that encompasses a Stacked Buck for the interface with an electrolyser, a Half-Bridge to interface with the ESS, and a Boost converter to interface with the RES.

These files are being constantly updated and at the moment (24th April 2025) these are the most updated ones.